Major changes. Variable IP, not only localhost! Normalisation of types: IP is now String(not integer), fileID is Integer (there were mixed Integer and Long issues). Changed and regenerated rso.thrift normalising and adding new server methods/changing interface for example to include slaveIP in request. Changes in DAO due to normalisation.Modified script not to insert test files. New client interface: put and get take 2 params: source and destination. Client now have variable masterIP given in commandline param as compared to earlier hardcoded one. DFSClosingClient for try-allocate resources java idiom.

+++ dfs-server/src/main/java/rso/dfs/model/File.java
@@ -5,7 +5,7 @@ package rso.dfs.model;
  * */
 public class File {
 
-	private Long id;
+	private Integer id;
 
 	private String name;
 
@@ -13,11 +13,11 @@ public class File {
 
 	private FileStatus status;
 
-	public long getId() {
+	public int getId() {
 		return id;
 	}
 
-	public void setId(long id) {
+	public void setId(int id) {
 		this.id = id;
 	}
